Issue #745 separated the streaming mesh node out of the multipart optimiser
