+++ bouncer/src/repo/core/model/bson/CMakeLists.txt
@@ -25,6 +25,7 @@ set(SOURCES
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_node_model_revision.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_node_reference.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_node_revision.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/repo_node_streaming_mesh.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_node_supermesh.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_node_texture.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_node_transformation.cpp
@@ -51,6 +52,7 @@ set(HEADERS
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_node_model_revision.h
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_node_reference.h
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_node_revision.h
+	${CMAKE_CURRENT_SOURCE_DIR}/repo_node_streaming_mesh.h
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_node_supermesh.h
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_node_texture.h
 	${CMAKE_CURRENT_SOURCE_DIR}/repo_node_transformation.h
