+++ bouncer/src/repo/manipulator/modeloptimizer/repo_optimizer_multipart.h
@@ -32,13 +32,14 @@
 #include "bvh/sweep_sah_builder.hpp"
 #include <repo/manipulator/modelconvertor/export/repo_model_export_abstract.h>
 #include <repo/core/model/bson/repo_bson.h>
+#include <repo/core/model/bson/repo_node_streaming_mesh.h>
 
 namespace repo {
 	namespace manipulator {
 		namespace modeloptimizer {
 			class MultipartOptimizer
 			{
-
+				
 				typedef float Scalar;
 				typedef bvh::Bvh<Scalar> Bvh;
 				typedef bvh::Vector3<Scalar> BvhVector3;
@@ -75,322 +76,8 @@ namespace repo {
 					bool isTexturedJob() const {
 						return !texId.isDefaultValue();
 					}
-				};
-								
-				class StreamingMeshNode {
-
-					class SupermeshingData {
-
-						repo::lib::RepoUUID uniqueId;
-
-						// Geometry
-						std::vector<repo::lib::RepoVector3D> vertices;
-						std::vector<repo::lib::repo_face_t> faces;
-						std::vector<repo::lib::RepoVector3D> normals;
-						std::vector<std::vector<repo::lib::RepoVector2D>> channels;
-
-					public:
-						SupermeshingData(
-							const repo::core::model::RepoBSON& bson,
-							const std::vector<uint8_t>& buffer,
-							const bool ignoreUVs)
-						{
-
-							this->uniqueId = bson.getUUIDField(REPO_NODE_LABEL_ID);
-
-							deserialise(bson, buffer, ignoreUVs);
-						}
-
-						repo::lib::RepoUUID getUniqueId() const {
-							return uniqueId;
-						}
-
-						std::uint32_t getNumFaces() const {
-							return faces.size();
-						}
-						std::vector<repo::lib::repo_face_t>& getFaces()
-						{
-							return faces;
-						}
-
-						std::uint32_t getNumVertices() const {
-							return vertices.size();
-						}
-						const std::vector<repo::lib::RepoVector3D>& getVertices() const {
-							return vertices;
-						}
-
-						void bakeMeshes(const repo::lib::RepoMatrix& transform) {
-							for (int i = 0; i < vertices.size(); i++) {
-								vertices[i] = transform * vertices[i];
-							}
-
-							for (int i = 0; i < normals.size(); i++) {
-								normals[i] = transform * normals[i];
-							}
-						}
-
-						const std::vector<repo::lib::RepoVector3D>& getNormals() const
-						{
-							return normals;
-						}
-
-						const std::vector<std::vector<repo::lib::RepoVector2D>>& getUVChannelsSeparated() const
-						{
-							return channels;
-						}
-
-					private:
-						void deserialise(
-							const repo::core::model::RepoBSON& bson,
-							const std::vector<uint8_t>& buffer,
-							const bool ignoreUVs)
-						{
-							auto blobRefBson = bson.getObjectField(REPO_LABEL_BINARY_REFERENCE);
-							auto elementsBson = blobRefBson.getObjectField(REPO_LABEL_BINARY_ELEMENTS);
-
-							if (elementsBson.hasField(REPO_NODE_MESH_LABEL_VERTICES)) {
-								auto vertBson = elementsBson.getObjectField(REPO_NODE_MESH_LABEL_VERTICES);
-								deserialiseVector(vertBson, buffer, vertices);
-							}
-
-							if (elementsBson.hasField(REPO_NODE_MESH_LABEL_NORMALS)) {
-								auto normBson = elementsBson.getObjectField(REPO_NODE_MESH_LABEL_NORMALS);
-								deserialiseVector(normBson, buffer, normals);
-							}
-
-							if (elementsBson.hasField(REPO_NODE_MESH_LABEL_FACES)) {
-
-								int32_t faceCount = bson.getIntField(REPO_NODE_MESH_LABEL_FACES_COUNT);
-								faces.reserve(faceCount);
-
-								std::vector<uint32_t> serialisedFaces = std::vector<uint32_t>();
-								auto faceBson = elementsBson.getObjectField(REPO_NODE_MESH_LABEL_FACES);
-								deserialiseVector(faceBson, buffer, serialisedFaces);
-
-								// Retrieve numbers of vertices for each face and subsequent
-								// indices into the vertex array.
-								// In API level 1, mesh is represented as
-								// [n1, v1, v2, ..., n2, v1, v2...]
-
-								int mNumIndicesIndex = 0;
-								while (serialisedFaces.size() > mNumIndicesIndex)
-								{
-									int mNumIndices = serialisedFaces[mNumIndicesIndex];
-									if (serialisedFaces.size() > mNumIndicesIndex + mNumIndices)
-									{
-										repo::lib::repo_face_t face;
-										face.resize(mNumIndices);
-										for (int i = 0; i < mNumIndices; ++i)
-											face[i] = serialisedFaces[mNumIndicesIndex + 1 + i];
-										faces.push_back(face);
-										mNumIndicesIndex += mNumIndices + 1;
-									}
-									else
-									{
-										repoError << "Cannot copy all faces. Buffer size is smaller than expected!";
-									}
-								}
-
-							}
-
-							if (!ignoreUVs && elementsBson.hasField(REPO_NODE_MESH_LABEL_UV_CHANNELS)) {
-								std::vector<repo::lib::RepoVector2D> serialisedChannels;
-								auto uvBson = elementsBson.getObjectField(REPO_NODE_MESH_LABEL_UV_CHANNELS);
-								deserialiseVector(uvBson, buffer, serialisedChannels);
-
-								if (serialisedChannels.size())
-								{
-									//get number of channels and split the serialised.
-									uint32_t nChannels = bson.getIntField(REPO_NODE_MESH_LABEL_UV_CHANNELS_COUNT);
-									uint32_t vecPerChannel = serialisedChannels.size() / nChannels;
-									channels.reserve(nChannels);
-									for (uint32_t i = 0; i < nChannels; i++)
-									{
-										channels.push_back(std::vector<repo::lib::RepoVector2D>());
-										channels[i].reserve(vecPerChannel);
-
-										uint32_t offset = i * vecPerChannel;
-										channels[i].insert(channels[i].begin(), serialisedChannels.begin() + offset,
-											serialisedChannels.begin() + offset + vecPerChannel);
-									}
-								}
-							}
-						}
-
-						template <class T>
-						void deserialiseVector(
-							const repo::core::model::RepoBSON& bson,
-							const std::vector<uint8_t>& buffer,
-							std::vector<T>& vec)
-						{
-							auto start = bson.getIntField(REPO_LABEL_BINARY_START);
-							auto size = bson.getIntField(REPO_LABEL_BINARY_SIZE);
-
-							vec.resize(size / sizeof(T));
-							memcpy(vec.data(), buffer.data() + (sizeof(uint8_t) * start), size);
-						}
-					};
-
-					repo::lib::RepoUUID sharedId;
-					std::uint32_t numVertices = 0;
-					repo::lib::RepoUUID parent;
-					repo::lib::RepoBounds bounds;
-
-					std::unique_ptr<SupermeshingData> supermeshingData;
-
-				public:
-					StreamingMeshNode()
-					{
-						// Default constructor so instances can be initialised for vectors
-					}
-
-					StreamingMeshNode(const repo::core::model::RepoBSON& bson)
-					{
-						if (bson.hasField(REPO_NODE_LABEL_SHARED_ID)) {
-							sharedId = bson.getUUIDField(REPO_NODE_LABEL_SHARED_ID);
-						}
-						if (bson.hasField(REPO_NODE_MESH_LABEL_VERTICES_COUNT)) {
-							numVertices = bson.getIntField(REPO_NODE_MESH_LABEL_VERTICES_COUNT);
-						}
-						if (bson.hasField(REPO_NODE_LABEL_PARENTS)) {
-							auto parents = bson.getUUIDFieldArray(REPO_NODE_LABEL_PARENTS);
-							parent = parents[0];
-						}
-						if (bson.hasField(REPO_NODE_MESH_LABEL_BOUNDING_BOX)) {
-							bounds = bson.getBoundsField(REPO_NODE_MESH_LABEL_BOUNDING_BOX);
-						}
-					}
-
-					bool supermeshingDataLoaded() {
-						return supermeshingData != nullptr;
-					}
-
-					void loadSupermeshingData(
-						const repo::core::model::RepoBSON &bson,
-						const std::vector<uint8_t>& buffer,
-						const bool ignoreUVs) {
-						if (supermeshingDataLoaded())
-						{							
-							repoWarning << "StreamingMeshNode instructed to load geometry data, but geometry data is already loaded.";
-							unloadSupermeshingData();
-						}
-
-						supermeshingData = std::make_unique<SupermeshingData>(bson, buffer, ignoreUVs);
-					}
-
-					void unloadSupermeshingData() {
-						supermeshingData.reset();
-					}
-
-					const repo::lib::RepoUUID getSharedId() const {
-						return sharedId;
-					}
-
-					const std::uint32_t getNumVertices() const
-					{
-						return numVertices;
-					}
-
-					const repo::lib::RepoBounds getBoundingBox() const
-					{
-						return bounds;
-					}
-
-					const repo::lib::RepoUUID getParent() const
-					{
-						return parent;
-					}
-
-					void transformBounds(const repo::lib::RepoMatrix &transform) {
-						auto newMinBound = transform * bounds.min();
-						auto newMaxBound = transform * bounds.max();
-						bounds = repo::lib::RepoBounds(newMinBound, newMaxBound);
-					}
-
-					// Requiring the supermeshing data to be loaded
-
-					const repo::lib::RepoUUID getUniqueId() {
-						if (supermeshingDataLoaded()) {
-							return supermeshingData->getUniqueId();
-						}
-						else {
-							repoError << "Tried to access supermesh geometry of StreamingMeshNode without loading geometry first. Empty returned.";
-							return repo::lib::RepoUUID();
-						}
-					}
-
-					const std::uint32_t getNumFaces() {
-						if (supermeshingDataLoaded()) {
-							return supermeshingData->getNumFaces();
-						}
-						else {
-							repoError << "Tried to access supermesh geometry of StreamingMeshNode without loading geometry first. Empty returned.";
-							return 0;
-						}
-					}
-
-					const std::vector<repo::lib::repo_face_t>& getFaces()
-					{
-						if (supermeshingDataLoaded()) {
-							return supermeshingData->getFaces();
-						}
-						else {
-							repoError << "Tried to access supermesh geometry of StreamingMeshNode without loading geometry first. Empty returned.";
-							return std::vector<repo::lib::repo_face_t>();
-						}
-					}
-
-					const std::uint32_t getNumVertices() {
-						if (supermeshingDataLoaded()) {
-							return supermeshingData->getNumVertices();
-						}
-						else {							
-							repoError << "Tried to access supermesh geometry of StreamingMeshNode without loading geometry first. Empty returned.";
-							return 0;
-						}
-					}
-					const std::vector<repo::lib::RepoVector3D>& getVertices() {
-						if (supermeshingDataLoaded()) {
-							return supermeshingData->getVertices();
-						}
-						else {
-							repoError << "Tried to access supermesh geometry of StreamingMeshNode without loading geometry first. Empty returned.";
-							return std::vector < repo::lib::RepoVector3D>();
-						}
-					}
-
-					void bakeMeshes(const repo::lib::RepoMatrix &transform) {
-						if (supermeshingDataLoaded()) {
-							return supermeshingData->bakeMeshes(transform);
-						}
-						else {							
-							repoError << "Tried to access supermesh geometry of StreamingMeshNode without loading geometry first. No action performed.";
-						}
-					}
-
-					const std::vector<repo::lib::RepoVector3D>& getNormals()
-					{
-						if (supermeshingDataLoaded()) {
-							return supermeshingData->getNormals();
-						}
-						else {
-							repoError << "Tried to access supermesh geometry of StreamingMeshNode without loading geometry first. Empty returned.";
-							return std::vector<repo::lib::RepoVector3D>();
-						}
-					}
-
-					const std::vector<std::vector<repo::lib::RepoVector2D>>& getUVChannelsSeparated()
-					{
-						if (supermeshingDataLoaded()) {
-							return supermeshingData->getUVChannelsSeparated();
-						}
-						else {
-							repoError << "Tried to access supermesh geometry of StreamingMeshNode without loading geometry first. Empty returned.";
-							return std::vector<std::vector<repo::lib::RepoVector2D>>();
-						}
-					}
-				};
+				};							
+				
 
 				typedef std::unordered_map <repo::lib::RepoUUID, std::shared_ptr<repo::core::model::MaterialNode>, repo::lib::RepoUUIDHasher> MaterialPropMap;
 				typedef std::unordered_map<repo::lib::RepoUUID, repo::lib::RepoMatrix, repo::lib::RepoUUIDHasher> TransformMap;
@@ -462,7 +149,7 @@ namespace repo {
 					repo::manipulator::modelconvertor::AbstractModelExport *exporter,
 					const TransformMap& transformMap,
 					const MaterialPropMap& matPropMap,
-					std::vector<StreamingMeshNode>& meshNodes,
+					std::vector<repo::core::model::StreamingMeshNode>& meshNodes,
 					const std::vector<std::vector<int>>& clusters,
 					const repo::lib::RepoUUID &texId
 				);
@@ -473,14 +160,14 @@ namespace repo {
 				);
 
 				void appendMesh(					
-					StreamingMeshNode &node,
+					repo::core::model::StreamingMeshNode &node,
 					const MaterialPropMap &matPropMap,
 					mapped_mesh_t &mappedMesh,
 					const repo::lib::RepoUUID &texId
 				);
 
 				Bvh buildFacesBvh(
-					StreamingMeshNode &node
+					repo::core::model::StreamingMeshNode &node
 				);
 
 				void flattenBvh(
@@ -504,7 +191,7 @@ namespace repo {
 				* each mapped_mesh_t has a vertex count below a certain size.
 				*/
 				void splitMesh(
-					StreamingMeshNode &node,
+					repo::core::model::StreamingMeshNode &node,
 					repo::manipulator::modelconvertor::AbstractModelExport *exporter,
 					const MaterialPropMap &matPropMap,
 					const repo::lib::RepoUUID &texId
@@ -523,23 +210,23 @@ namespace repo {
 				* geometry size.
 				*/
 				std::vector<std::vector<int>> clusterMeshNodes(
-					const std::vector<StreamingMeshNode>& nodes
+					const std::vector<repo::core::model::StreamingMeshNode>& nodes
 				);
 
 				void clusterMeshNodesBvh(
-					const std::vector<StreamingMeshNode>& meshes,
+					const std::vector<repo::core::model::StreamingMeshNode>& meshes,
 					const std::vector<int>& binIndexes,
 					std::vector<std::vector<int>>& clusters);
 
 				Bvh buildBoundsBvh(
 					const std::vector<int>& binIndexes,
-					const std::vector<StreamingMeshNode>& meshes
+					const std::vector<repo::core::model::StreamingMeshNode>& meshes
 				);
 
 				std::vector<size_t> getVertexCounts(
 					const Bvh& bvh,
 					const std::vector<int>& binIndexes,
-					const std::vector<StreamingMeshNode>& meshes
+					const std::vector<repo::core::model::StreamingMeshNode>& meshes
 				);
 
 				std::vector<size_t> getSupermeshBranchNodes(
